✨ feat(api): add express app
Create Express App

Create Express App at server.js. Add middleware

Add auth middleware to authenticate with JWT and actor router

auth middleware is used to verify the jwt of the request to set the user
and the actor router to perform BREAD operations to the actor table of
the database.